move print function to common place
